+++ app/vmselect/netstorage/netstorage.go
@@ -2494,7 +2494,6 @@ func (sn *storageNode) processSearchQuery(qt *querytracer.Tracer, requestData []
 
 func (sn *storageNode) execOnConnWithPossibleRetry(qt *querytracer.Tracer, funcName string, f func(bc *handshake.BufferedConn) error, deadline searchutil.Deadline) error {
 	qtChild := qt.NewChild("rpc call %s()", funcName)
-	logger.Infof("jayice first try")
 	err := sn.execOnConn(qtChild, funcName, f, deadline)
 	defer qtChild.Done()
 	if err == nil {
@@ -2504,7 +2503,6 @@ func (sn *storageNode) execOnConnWithPossibleRetry(qt *querytracer.Tracer, funcN
 	var ne net.Error
 	var le *limitExceededErr
 	if errors.As(err, &le) || errors.As(err, &er) || errors.As(err, &ne) && ne.Timeout() || deadline.Exceeded() || errors.Is(err, errCannotObtainConn) {
-		logger.Infof("jayice dont retry")
 		// There is no sense in repeating the query on the following errors:
 		//
 		//   - exceeded complexity limits (limitExceededErr)
@@ -2517,7 +2515,6 @@ func (sn *storageNode) execOnConnWithPossibleRetry(qt *querytracer.Tracer, funcN
 	println(err.Error())
 	// Repeat the query in the hope the error was temporary.
 	qtRetry := qtChild.NewChild("retry rpc call %s() after error", funcName)
-	logger.Infof("jayice second try")
 
 	err = sn.execOnConn(qtRetry, funcName, f, deadline)
 	qtRetry.Done()
@@ -2537,7 +2534,7 @@ func (sn *storageNode) execOnConn(qt *querytracer.Tracer, funcName string, f fun
 	if timeout <= 0 {
 		return fmt.Errorf("request timeout reached: %s", deadline.String())
 	}
-	bc, err := sn.connPool.GetV2()
+	bc, err := sn.connPool.Get()
 	if err != nil {
 		return fmt.Errorf("%w: %w", errCannotObtainConn, err)
 	}
@@ -3262,7 +3259,7 @@ func newStorageNode(ms *metrics.Set, group *storageNodesGroup, addr string) *sto
 	addr = normalizedAddr
 
 	// There is no need in requests compression, since vmselect requests are usually very small.
-	connPool := netutil.NewConnPool(ms, "vmselect", addr, handshake.VMSelectClient, 0, *vmstorageDialTimeout, *vmstorageUserTimeout)
+	connPool := netutil.NewConnPool(ms, "vmselect", addr, handshake.VMSelectClient, 0, *vmstorageDialTimeout, *vmstorageUserTimeout, handshake.HealthCheckToVmstroage)
 
 	sn := &storageNode{
 		group:    group,