+++ lib/netutil/conn_pool.go
@@ -2,11 +2,9 @@ package netutil
 
 import (
 	"fmt"
-	"io"
 	"sync"
 	"time"
 
-	"github.com/VictoriaMetrics/VictoriaMetrics/lib/encoding"
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/fasttime"
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/handshake"
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/logger"
@@ -24,9 +22,12 @@ type ConnPool struct {
 	// See https://github.com/VictoriaMetrics/VictoriaMetrics/issues/2552
 	concurrentDialsCh chan struct{}
 
-	name             string
-	handshakeFunc    handshake.Func
-	compressionLevel int
+	name                  string
+	handshakeFunc         handshake.Func
+	healthCheckFunc       handshake.HealthCheckFunc
+	compressionLevel      int
+	connKeepAliveInterval uint64
+	dialTimeout           time.Duration
 
 	conns []connWithTimestamp
 
@@ -45,7 +46,8 @@ type connWithTimestamp struct {
 }
 
 var (
-	concurrentDialLimit = 8
+	concurrentDialLimit                 = 8
+	defaultConnKeepAliveInterval uint64 = 10
 )
 
 // InitConcurrentDialLimit initiates the concurrentDialLimit with value between 8 and 64 based on the given concurrentRequestLimit.
@@ -64,17 +66,19 @@ func InitConcurrentDialLimit(concurrentRequestLimit int) {
 // The compression is disabled if compressionLevel <= 0.
 //
 // Call ConnPool.MustStop when the returned ConnPool is no longer needed.
-func NewConnPool(ms *metrics.Set, name, addr string, handshakeFunc handshake.Func, compressionLevel int, dialTimeout, userTimeout time.Duration) *ConnPool {
+func NewConnPool(ms *metrics.Set, name, addr string, handshakeFunc handshake.Func, compressionLevel int, dialTimeout, userTimeout time.Duration, healthCheckFunc handshake.HealthCheckFunc) *ConnPool {
 	cp := &ConnPool{
 		d:                 NewTCPDialer(ms, name, addr, dialTimeout, userTimeout),
 		concurrentDialsCh: make(chan struct{}, concurrentDialLimit),
 
-		name:             name,
-		handshakeFunc:    handshakeFunc,
-		compressionLevel: compressionLevel,
+		name:                  name,
+		handshakeFunc:         handshakeFunc,
+		healthCheckFunc:       healthCheckFunc,
+		compressionLevel:      compressionLevel,
+		connKeepAliveInterval: defaultConnKeepAliveInterval,
+		dialTimeout:           dialTimeout,
 	}
-	cp.checkAvailability(true)
-	cp.checkAvailability(true)
+	//cp.checkAvailability(true)
 	_ = ms.NewGauge(fmt.Sprintf(`vm_tcpdialer_conns_idle{name=%q, addr=%q}`, name, addr), func() float64 {
 		cp.mu.Lock()
 		n := len(cp.conns)
@@ -135,100 +139,41 @@ func (cp *ConnPool) Addr() string {
 	return cp.d.addr
 }
 
-// Get returns free connection from the pool.
 func (cp *ConnPool) Get() (*handshake.BufferedConn, error) {
 	c, err := cp.tryGetConn()
 	if err != nil {
 		return nil, err
 	}
 	if c.bc != nil {
-		// Fast path - obtained the connection from pool.
-		return c.bc, nil
-	}
-	return cp.getConnSlow()
-}
-
-func (cp *ConnPool) healthCheck(bc *handshake.BufferedConn) bool {
-	logger.Infof("jayice health check")
-	time.Sleep(6 * time.Second)
-
-	funcName := "healthCheck_v1"
-	buf := []byte(funcName)
-	sizeBuf := encoding.MarshalUint64(nil, uint64(len(buf)))
-	if _, err := bc.Write(sizeBuf); err != nil {
-		return false
-	}
-	_, err := bc.Write(buf)
-	if err != nil {
-		return false
-	}
-	var trace [1]byte
-	_, err = bc.Write(trace[:])
-	if err != nil {
-		return false
-	}
-	timeout := encoding.MarshalUint32(nil, 5)
-	_, err = bc.Write(timeout)
-	if err != nil {
-		return false
-	}
-	if err = bc.Flush(); err != nil {
-		return false
-	}
-
-	var resp [8]byte
-	if _, err = io.ReadFull(bc, resp[:]); err != nil {
-		return false
-	}
-	n := encoding.UnmarshalUint64(resp[:])
-	logger.Infof("jayice health check success:%d", n)
-
-	if _, err = io.ReadFull(bc, resp[:]); err != nil {
-		return false
-	}
-	n = encoding.UnmarshalUint64(resp[:])
-	logger.Infof("jayice health check trace size:%d", n)
-
-	return true
-}
-
-func (cp *ConnPool) GetV2() (*handshake.BufferedConn, error) {
-	c, err := cp.tryGetConn()
-	if err == nil && c.bc != nil {
-		// fast path: fresh connection don't need health check
-		if fasttime.UnixTimestamp()-c.lastActiveTime < 30000 {
-			logger.Infof("jayice1")
+		// fast path: fresh connection can be returned without health check
+		if fasttime.UnixTimestamp()-c.lastActiveTime < cp.connKeepAliveInterval {
 			return c.bc, nil
 		}
 	} else {
-		//
+		// slow path: no connections in the pool, dial a new connection.
 		return cp.getConnSlow()
 	}
 
-	// slow path
-	timeout := time.NewTimer(5 * time.Second)
+	// slow path: perform health check for the existing connection.
+	timeout := time.NewTimer(cp.dialTimeout + 1*time.Second)
 	defer timeout.Stop()
 	connChannel := make(chan *handshake.BufferedConn, 1)
 	done := make(chan struct{})
 	defer close(done)
+
 	go func() {
 		defer close(connChannel)
-		if c.bc != nil {
-			logger.Infof("jayice4")
-			if cp.healthCheck(c.bc) {
-				select {
-				case connChannel <- c.bc:
-				case <-done:
-					logger.Infof("jayice iam timeout")
-					cp.Put(c.bc)
-				}
-				return
-			} else {
-				logger.Infof("jayice health check failed")
-				_ = c.bc.Close()
-				c.bc = nil
+		if err = cp.healthCheckFunc(c.bc); err == nil {
+			select {
+			case connChannel <- c.bc:
+			case <-done:
+				cp.Put(c.bc)
 			}
+			return
 		}
+		_ = c.bc.Close()
+		c.bc = nil
+
 		for {
 			select {
 			case <-done:
@@ -238,7 +183,6 @@ func (cp *ConnPool) GetV2() (*handshake.BufferedConn, error) {
 				return
 			default:
 			}
-			logger.Infof("jayice test1")
 
 			c, err = cp.tryGetConn()
 			if err != nil || c.bc == nil {
@@ -248,7 +192,7 @@ func (cp *ConnPool) GetV2() (*handshake.BufferedConn, error) {
 				}
 				return
 			}
-			if fasttime.UnixTimestamp()-c.lastActiveTime < 30000 || cp.healthCheck(c.bc) {
+			if fasttime.UnixTimestamp()-c.lastActiveTime < cp.connKeepAliveInterval || cp.healthCheckFunc(c.bc) == nil {
 				select {
 				case connChannel <- c.bc:
 				case <-done:
@@ -256,6 +200,7 @@ func (cp *ConnPool) GetV2() (*handshake.BufferedConn, error) {
 				}
 				return
 			} else {
+				time.Sleep(100 * time.Millisecond)
 				_ = c.bc.Close()
 				c.bc = nil
 				continue
@@ -265,14 +210,13 @@ func (cp *ConnPool) GetV2() (*handshake.BufferedConn, error) {
 
 	select {
 	case conn := <-connChannel:
-		logger.Infof("jayice2")
 		if conn != nil {
 			return conn, nil
 		}
+		// no more connections in the pool, dial a new connection.
 		return cp.getConnSlow()
 	case <-timeout.C:
-		logger.Infof("jayice3")
-		return nil, fmt.Errorf("error")
+		return nil, fmt.Errorf("failed to get valid connection to %s within 5 seconds", cp.d.addr)
 	}
 }
 
@@ -334,7 +278,6 @@ func (cp *ConnPool) tryGetConn() (connWithTimestamp, error) {
 	if len(cp.conns) == 0 {
 		return connWithTimestamp{}, cp.lastDialError
 	}
-	logger.Infof("jayice con len:%d", len(cp.conns))
 	c := cp.conns[len(cp.conns)-1]
 	cp.conns = cp.conns[:len(cp.conns)-1]
 	return c, nil
@@ -344,7 +287,6 @@ func (cp *ConnPool) tryGetConn() (connWithTimestamp, error) {
 //
 // Do not put broken and closed connections to the pool!
 func (cp *ConnPool) Put(bc *handshake.BufferedConn) {
-	logger.Infof("jayice put")
 	if err := bc.SetDeadline(time.Time{}); err != nil {
 		// Close the connection instead of returning it to the pool,
 		// since it may be broken.